requirements.txt
django>=6.0.1
django-crispy-forms>=2.5
django-environ>=0.12.0
django-filter>=25.2
djangorestframework>=3.16.1
psycopg2-binary>=2.9.11
requests>=2.32.5

# Development and testing dependencies
black>=26.1.0  # code formatter
ruff>=0.14.14  # linter
pytest>=9.0.2  # testing framework
pytest-django>=4.11.1  # Django plugin for pytest
